+++ packages/js-sdk/src/clients/spclient/auth.ts
@@ -1,6 +1,7 @@
 import { AuthType, ReqMeta } from '@/types/auth';
 import { hexlify, joinSignature } from '@ethersproject/bytes';
 import { SigningKey } from '@ethersproject/signing-key';
+import { toUtf8Bytes } from '@ethersproject/strings';
 import { ed25519 } from '@noble/curves/ed25519';
 import { Headers } from 'cross-fetch';
 import { keccak256 } from 'ethereum-cryptography/keccak.js';
@@ -196,16 +197,14 @@ export const encodePath = (pathName: string) => {
         continue;
 
       // others characters need to be encoded
-      default:
-        // . ! @ # $ % ^ & * ) ( - + = { } [ ] / " , ' < > ~ \ .` ? : ; | \\
-        if (/[.!@#\$%\^&\*\)\(\-+=\{\}\[\]\/\",'<>~\·`\?:;|\\]+$/.test(s)) {
-          // english characters
-          const hexStr = s.charCodeAt(0).toString(16);
-          encodedPathName += '%' + hexStr.toUpperCase();
-        } else {
-          // others characters
-          encodedPathName += encodeURI(s);
+      default: {
+        const u = toUtf8Bytes(s);
+
+        for (let i = 0; i < u.length; i++) {
+          const hexStr = hexlify(u[i]);
+          encodedPathName += '%' + hexStr.slice(2).toUpperCase();
         }
+      }
     }
   }
   return encodedPathName;