+++ packages/bsc-cross-sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @bnb-chain/bsc-cross-greenfield-sdk
 
+## 0.0.3
+
+### Patch Changes
+
+- [#551](https://github.com/bnb-chain/greenfield-js-sdk/pull/551)
+  [`acdf20f879f21140fc355ca38c532d523263076d`](https://github.com/bnb-chain/greenfield-js-sdk/commit/acdf20f879f21140fc355ca38c532d523263076d)
+  Thanks [@rrr523](https://github.com/rrr523)! - feat: export `encodePath`
+
+## 0.0.3-alpha.0
+
+### Patch Changes
+
+- [#551](https://github.com/bnb-chain/greenfield-js-sdk/pull/551)
+  [`acdf20f879f21140fc355ca38c532d523263076d`](https://github.com/bnb-chain/greenfield-js-sdk/commit/acdf20f879f21140fc355ca38c532d523263076d)
+  Thanks [@rrr523](https://github.com/rrr523)! - feat: export `encodePath`
+
 ## 0.0.2
 
 ### Patch Changes